+++ minidenticons.js
@@ -15,13 +15,13 @@ function simpleHash(str) {
 }
 
 /**
- * @type {import('.').identicon}
+ * @type {import('.').minidenticon}
  */
-export function identicon(username="", saturation=DEFAULT_SATURATION, lightness=DEFAULT_LIGHTNESS) {
-    const hash = simpleHash(username)
+export function minidenticon(seed="", saturation=DEFAULT_SATURATION, lightness=DEFAULT_LIGHTNESS) {
+    const hash = simpleHash(seed)
     // console.log("%c" + hash.toString(2).padStart(32, "0"), "font-family:monospace") // uncomment to debug
     const hue = (hash % COLORS_NB) * (360 / COLORS_NB)
-    return [...Array(username ? 25 : 0)].reduce((acc, e, i) =>
+    return [...Array(seed ? 25 : 0)].reduce((acc, e, i) =>
         // testing the 15 lowest weight bits of the hash
         hash & (1 << (i % 15)) ?
             acc + `<rect x="${i > 14 ? 7 - ~~(i / 5) : ~~(i / 5)}" y="${i % 5}" width="1" height="1"/>`
@@ -35,10 +35,10 @@ export function identicon(username="", saturation=DEFAULT_SATURATION, lightness=
 /**
  * @type {void}
  */
-export const identiconSvg =
+export const minidenticonSvg =
     // declared as a pure function to be tree-shaken by the bundler
-    /*@__PURE__*/globalThis.customElements?.define('identicon-svg',
-        class IdenticonSvg extends HTMLElement {
+    /*@__PURE__*/globalThis.customElements?.define('minidenticon-svg',
+        class MinidenticonSvg extends HTMLElement {
             static observedAttributes = ['username', 'saturation', 'lightness']
             // private fields to allow Terser mangling
             static #memoized = {}
@@ -50,12 +50,12 @@ export const identiconSvg =
             // attributeChangedCallback() is called for every observed attribute before connectedCallback()
             attributeChangedCallback() { if (this.#isConnected) this.#setContent() }
             #setContent() {
-                const args = IdenticonSvg.observedAttributes
+                const args = MinidenticonSvg.observedAttributes
                                 .map(key => this.getAttribute(key) || undefined)
                 const memoKey = args.join(',')
-                this.innerHTML = IdenticonSvg.#memoized[memoKey] ??=
+                this.innerHTML = MinidenticonSvg.#memoized[memoKey] ??=
                     // @ts-ignore
-                    identicon(...args)
+                    minidenticon(...args)
             }
         }
     )